build(initial)!: First version of the code
This package will use a redis-like cache store to set "lock files"
atomically such that it can be used in a distributed environment
when multiple things are competing for the same resource. The
TTL and refresh of the lock is configurable as well as the contents
of the lock for use with meta data (such as server/pod name).

Fixes: #1