chore(lib)!: Change the return value for `inspect`
The `inspect` function would be more helpful if it also
returned the TTL as well as the contents. Also, do NOT
require the lock to be held in order to inspect. This way,
other clients can view the contents if there is useful
metadata stored in the lock.

BREAKING CHANGE: The function now returns a staic object with
properties for TTL and contents

BREAKING CHANGE: `acquire` is no longer required to `inspect`

Fixes: #6